refactor: move the network actions to actions.ts
